backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
django-otp>=1.4.0
pyotp>=2.9.0

# File Handling
openpyxl>=3.1.2
pandas>=2.1.0

# Email & SMS
django-sendgrid-v5>=0.9.0
twilio>=9.0.0

# Notifications
firebase-admin>=6.4.0

# Payment Gateway
razorpay>=1.3.0
cashfree-pg>=1.0.0

# Utilities
python-dotenv>=1.0.0
django-redis>=5.4.0
celery>=5.3.0
django-celery-beat>=2.5.0

# Security
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0

# Documentation
drf-yasg>=1.21.7

gunicorn
whitenoise